core: Adds useful debugging macros and error enum for better communication of errors for error handling.
This addition of a new error enum type will allow more readable error handling. It will also be more useful than just returning a 1 or 0 from functions to determine success or failure. Now we can selectively fail if a type of error can't be resolved. This change moves core to use the ENGINE_ERROR enum to return errors from functions.