Functions that needed to return the success of their operation now use ENGINE_ERROR to communitcate it.
Originally the success of functions was indicated by using a single integer of 1 or 0. This wasn't descriptive and you couldn't tell what went wrong from outside. Now with the use of ENGINE_ERROR, helpful and specific error handling can be used to keep the application running or abort it, if it is too far gone.